Update
I formatted the numbers being displayed to default to 2-digit (00) format. I think it looks more aesthetically pleasing. I also put a border around the total time, made the font larger and resized the inputs. I got rid of the H1 element on mobile, but not sure why it reappeared, but I took care of that as well.